feat: Update npm dependencies to include "canvas-confetti" library
This commit updates the npm dependencies in the package.json and pnpm-lock.yaml files to include the "canvas-confetti" library. It also adds the corresponding devDependency in the devDependencies section. This update is necessary to introduce the "canvas-confetti" library into the project, which will be used to create confetti animations in the StartScreen component when the save button is clicked. The library will enhance the visual appeal and interactivity of the component, providing a more engaging user experience.